+++ demis/output.tf
@@ -15,22 +15,27 @@ output "stage_name" {
 }
 
 output "version_istio_routing_chart" {
-  value       = local.istio_routing_chart_version
+  value       = module.demis_services.version_istio_routing_chart
   description = "Version of the Istio Routing Chart being used"
 }
 
 output "version_fhir_profile_snapshots" {
-  value       = local.fhir_profile_snapshots
+  value       = module.demis_services.fhir_profile_snapshots
   description = "Version of the FHIR Profile Snapshots being used"
 }
 
 output "version_igs_profile_snapshots" {
-  value       = local.igs_profile_snapshots
+  value       = module.demis_services.igs_profile_snapshots
   description = "Version of the IGS Profile Snapshots being used"
 }
 
+output "ars_profile_snapshots" {
+  value       = module.demis_services.ars_profile_snapshots
+  description = "Version of the ARS Profile Snapshots being used"
+}
+
 output "version_routing_data" {
-  value       = local.routing_data_version
+  value       = module.demis_services.version_routing_data
   description = "Version of the Routing Data"
 }