+++ idm/cronjobs.tf
@@ -4,36 +4,31 @@ locals {
   kup_version = replace(module.idm_services.kup_version, ".", "-")
 }
 
-resource "null_resource" "cus_manual_trigger" {
-  count = module.idm_services.cus_enabled ? 1 : 0
+resource "terraform_data" "cus_manual_trigger" {
   provisioner "local-exec" {
     environment = {
       KUBECONFIG = var.kubeconfig_path
     }
-    command = "kubectl create job -n ${var.target_namespace} --from=cronjob/certificate-update-service-${local.cus_version} manual-cus-${substr(sha256(timestamp()), 0, 10)}"
+    command = "if [ ${module.idm_services.cus_enabled} = true ]; then kubectl create job -n ${var.target_namespace} --from=cronjob/certificate-update-service-${local.cus_version} manual-cus-${substr(sha256(timestamp()), 0, 10)}; fi"
   }
 
   # triggers the job when the deployment is executed
-  triggers = {
-    last_execution = timestamp()
-  }
+
+  triggers_replace = [timestamp()]
 
   depends_on = [module.idm_services]
 }
 
-resource "null_resource" "kup_manual_trigger" {
-  count = module.idm_services.kup_enabled ? 1 : 0
+resource "terraform_data" "kup_manual_trigger" {
   provisioner "local-exec" {
     environment = {
       KUBECONFIG = var.kubeconfig_path
     }
-    command = "kubectl create job -n ${var.target_namespace} --from=cronjob/keycloak-user-purger-${local.kup_version} manual-kup-${substr(sha256(timestamp()), 0, 10)}"
+    command = "if [ ${module.idm_services.kup_enabled} = true ]; then kubectl create job -n ${var.target_namespace} --from=cronjob/keycloak-user-purger-${local.kup_version} manual-kup-${substr(sha256(timestamp()), 0, 10)}; fi"
   }
 
   # triggers the job when the deployment is executed
-  triggers = {
-    last_execution = timestamp()
-  }
+  triggers_replace = [timestamp()]
 
   depends_on = [module.idm_services]
 }