+++ idm/providers.tf
@@ -8,12 +8,7 @@ terraform {
 
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "2.36.0"
-    }
-
-    null = {
-      source  = "hashicorp/null"
-      version = "3.2.3"
+      version = "2.37.1"
     }
   }
 }